recipe_project/requirements.txt
# =============================================================================
# Recipe Modification Extraction - Dependencies
# =============================================================================
# Group 11: Daniel Simanovsky & Roei Ben Artzi
# NLP 2025a - Tel Aviv University
#
# Install: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core ML/NLP
# -----------------------------------------------------------------------------
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
tokenizers>=0.15.0
accelerate>=0.25.0

# -----------------------------------------------------------------------------
# Evaluation
# -----------------------------------------------------------------------------
seqeval>=1.2.2
scikit-learn>=1.3.2
scipy>=1.11.4

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pandas>=2.1.4
numpy>=1.26.3
tqdm>=4.66.1

# -----------------------------------------------------------------------------
# APIs
# -----------------------------------------------------------------------------
# YouTube Data Collection
google-api-python-client>=2.100.0

# Google AI (Gemini) - Primary Teacher Model
google-generativeai>=0.4.0

# OpenAI (GPT-4o) - Backup Teacher Model
openai>=1.6.1

# -----------------------------------------------------------------------------
# Configuration & Utilities
# -----------------------------------------------------------------------------
PyYAML>=6.0
python-dotenv>=1.0.0
python-bidi>=0.4.2

# -----------------------------------------------------------------------------
# Visualization
# -----------------------------------------------------------------------------
matplotlib>=3.8.2
seaborn>=0.13.1

# -----------------------------------------------------------------------------
# Development (optional)
# -----------------------------------------------------------------------------
jupyter>=1.0.0
ipykernel>=6.28.0
tensorboard>=2.15.0
